requirements.txt
# Core ML and Data Science
tensorflow>=2.13.0
#tensorflow-gpu>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Asynpython monitor_positions.pyc and Performance
uvloop>=0.17.0
aiohttp>=3.8.0
aiofiles>=23.0.0

# Database and Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# redis>=4.5.0  # Опціонально: для multi-bot setup (розподілений кеш)
cachetools>=5.3.0  # Memory cache (використовується зараз)

# Data Processing
numba>=0.57.0
polars>=0.18.0 

# API and Web
requests>=2.31.0
python-binance>=1.0.16
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
textblob>=0.17.1

# Telegram Integration
python-telegram-bot>=20.0

# GPU Utilities
GPUtil>=1.4.0

# System Monitoring
psutil>=5.9.0
nvidia-ml-py3>=7.352.0 

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and Configuration
colorlog>=6.7.0
pyyaml>=6.0

# Optimization
cython>=3.0.0

# Rust Extensions (build)
maturin>=1.2.0  # Для збірки Rust модулів
setuptools-rust>=1.7.0

# Optional but recommended
jupyter>=1.0.0
jupyterlab>=4.0.0
plotly>=5.15.0
dash>=2.12.0  # Для веб-дашборду

# Production deployment
gunicorn>=21.0.0
uvicorn>=0.23.0
fastapi>=0.101.0  # Для REST API